Move tests to labeled dataset
